+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to
 
 - cosmwasm-vm: Add `secp256r1_verify` and `secp256r1_recover_pubkey` imports for
   ECDSA signature verification over secp256r1. ([#1983], [#2057], [#2058])
+- cosmwasm-vm: Add metrics for the pinned memory cache ([#2059])
 
 [#1983]: https://github.com/CosmWasm/cosmwasm/pull/1983
 [#2057]: https://github.com/CosmWasm/cosmwasm/pull/2057
@@ -23,6 +24,7 @@ and this project adheres to
 
 [#2044]: https://github.com/CosmWasm/cosmwasm/pull/2044
 [#2051]: https://github.com/CosmWasm/cosmwasm/pull/2051
+[#2059]: https://github.com/CosmWasm/cosmwasm/pull/2059
 
 ## [2.0.0] - 2024-03-12